backend/app/tests/unit/test_queue_and_limits.py
"""
Unit Tests for Queue Management and Limits.

Tests queue behavior, rate limits, and resource constraints.
"""

import pytest
from unittest.mock import Mock, patch
from datetime import datetime, timedelta

from app.core.limits import (
    LLMPurpose, ModelConfig, BudgetLimits, QueueLimits,
    get_model_config, calculate_cost
)


class TestModelConfig:
    """Test model configuration and limits."""

    def test_get_model_config_guardian(self):
        """Test getting config for guardian purpose."""
        config = get_model_config(LLMPurpose.GUARDIAN)

        assert config is not None
        assert config.model is not None
        assert config.temperature >= 0.0
        assert config.max_tokens_per_call > 0
        assert config.max_calls_per_task > 0

    def test_get_model_config_fix_generation(self):
        """Test getting config for fix generation."""
        config = get_model_config(LLMPurpose.FIX_GENERATION)

        assert config is not None
        assert config.model is not None

    def test_get_model_config_cit_generation(self):
        """Test getting config for CIT generation."""
        config = get_model_config(LLMPurpose.CIT_GENERATION)

        assert config is not None
        assert config.model is not None

    def test_all_purposes_have_config(self):
        """Test that all LLM purposes have configuration."""
        for purpose in LLMPurpose:
            config = get_model_config(purpose)
            assert config is not None, f"Missing config for {purpose}"


class TestCostCalculation:
    """Test cost calculation for different models."""

    def test_calculate_cost_basic(self):
        """Test basic cost calculation."""
        # Using a known model pricing
        cost = calculate_cost(
            model="gpt-4o-mini",
            prompt_tokens=1000,
            completion_tokens=500
        )

        # gpt-4o-mini: $0.15 per 1M input, $0.60 per 1M output
        expected = (1000 * 0.15 / 1_000_000) + (500 * 0.60 / 1_000_000)
        assert abs(cost - expected) < 0.0001

    def test_calculate_cost_zero_tokens(self):
        """Test cost calculation with zero tokens."""
        cost = calculate_cost(
            model="gpt-4o-mini",
            prompt_tokens=0,
            completion_tokens=0
        )

        assert cost == 0.0

    def test_calculate_cost_large_numbers(self):
        """Test cost calculation with large token counts."""
        cost = calculate_cost(
            model="gpt-4o-mini",
            prompt_tokens=100000,
            completion_tokens=50000
        )

        assert cost > 0
        assert cost < 100  # Sanity check


class TestBudgetLimits:
    """Test budget limit constants."""

    def test_budget_limits_exist(self):
        """Test that budget limits are defined."""
        assert BudgetLimits.MAX_TOKENS_PER_TASK > 0
        assert BudgetLimits.MAX_COST_PER_TASK_USD > 0
        assert BudgetLimits.MAX_COST_PER_USER_PER_DAY_USD > 0

    def test_budget_limits_reasonable(self):
        """Test that budget limits are reasonable values."""
        # Token limit should be in thousands
        assert BudgetLimits.MAX_TOKENS_PER_TASK >= 1000

        # Cost limits should be reasonable (not too low or too high)
        assert 0.1 <= BudgetLimits.MAX_COST_PER_TASK_USD <= 100
        assert 1.0 <= BudgetLimits.MAX_COST_PER_USER_PER_DAY_USD <= 1000


class TestQueueLimits:
    """Test queue limit constants."""

    def test_queue_limits_exist(self):
        """Test that queue limits are defined."""
        assert QueueLimits.MAX_QUEUE_SIZE > 0
        assert QueueLimits.MAX_TASKS_PER_USER_PER_DAY > 0

    def test_queue_limits_reasonable(self):
        """Test that queue limits are reasonable."""
        # Queue should hold at least a few tasks
        assert QueueLimits.MAX_QUEUE_SIZE >= 10

        # User shouldn't be limited too much
        assert QueueLimits.MAX_TASKS_PER_USER_PER_DAY >= 5


class TestPatchApplication:
    """Test patch application logic (placeholder for actual implementation)."""

    def test_patch_types_defined(self):
        """Test that patch types are well-defined."""
        # This is a placeholder - actual patch logic would be tested here
        patch_types = ["replace", "insert", "delete"]

        for patch_type in patch_types:
            assert patch_type in ["replace", "insert", "delete"]

    def test_patch_validation(self):
        """Test patch validation logic."""
        # Example patch structure from code_agent_v1.json
        valid_patch = {
            "file_path": "src/test.py",
            "patch_type": "replace",
            "start_line": 10,
            "end_line": 15,
            "new_code": "print('hello')",
            "description": "Fix print statement"
        }

        # All required fields present
        required_fields = ["file_path", "patch_type", "start_line", "description"]
        for field in required_fields:
            assert field in valid_patch


class TestRetryBehavior:
    """Test retry behavior for queue operations."""

    def test_transient_errors_retryable(self):
        """Test that transient errors are retried."""
        from app.core.errors import ErrorType, ERROR_TAXONOMY

        transient_types = [
            ErrorType.NETWORK_TIMEOUT,
            ErrorType.LLM_RATE_LIMIT,
            ErrorType.SANDBOX_TIMEOUT
        ]

        for error_type in transient_types:
            error_info = ERROR_TAXONOMY.get(error_type)
            assert error_info is not None
            assert error_info.retry_policy.should_retry is True

    def test_permanent_errors_not_retryable(self):
        """Test that permanent errors are not retried."""
        from app.core.errors import ErrorType, ERROR_TAXONOMY

        permanent_types = [
            ErrorType.FILE_NOT_FOUND,
            ErrorType.GUARDIAN_REJECTED,
            ErrorType.LLM_INVALID_RESPONSE
        ]

        for error_type in permanent_types:
            error_info = ERROR_TAXONOMY.get(error_type)
            assert error_info is not None
            assert error_info.retry_policy.should_retry is False


class TestTaskPriorities:
    """Test task prioritization logic (placeholder)."""

    def test_priority_levels(self):
        """Test that priority levels are defined."""
        # Placeholder for actual priority logic
        priorities = ["low", "normal", "high", "urgent"]

        assert len(priorities) > 0
        assert "normal" in priorities


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
