backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.29.0
sqlalchemy==2.0.23
pydantic>=2.9.0,<3.0.0
openai>=1.58.1,<2.0.0
python-dotenv>=1.0.0

# Codebase Guardian / SCM layer dependencies
tree-sitter==0.21.0
tree-sitter-python==0.21.0
sentence-transformers==2.2.2
chromadb==0.4.22
torch==2.1.2

# Git integration dependencies
GitPython==3.1.40
PyGithub==2.1.1

# Container/Docker dependencies
docker==7.0.0

# Real-time updates and WebSocket dependencies
websockets==12.0

# CIT Agent / E2E testing dependencies
playwright==1.40.0

# Observability and monitoring dependencies
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-instrumentation-fastapi>=0.43b0
opentelemetry-instrumentation-sqlalchemy>=0.43b0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-exporter-jaeger>=1.21.0
prometheus-client==0.19.0

# Background task queue dependencies
rq==1.16.0
redis==5.0.1

alembic